Begin separating PHP 5 and PHP 7 engine hooks
Increasingly it is becoming difficult to add new functionality
because of PHP 5. Adding PHP 7 specific features is also
difficult because of accidentally breaking PHP 5.
