+++ package.xml
@@ -87,6 +87,12 @@
                         <file name="mpack.c" role="src" />
                         <file name="mpack.h" role="src" />
                     </dir>
+                    <dir name="php5">
+                        <file name="engine_hooks.c" role="src" />
+                    </dir>
+                    <dir name="php7">
+                        <file name="engine_hooks.c" role="src" />
+                    </dir>
                     <dir name="third-party">
                         <file name="mt19937-64.c" role="src" />
                         <file name="mt19937-64.h" role="src" />
